stop delete categories when books are attached

+++ app/Repositories/CategoryRepository.php
@@ -5,6 +5,7 @@ namespace App\Repositories;
 use App\Models\Category;
 use App\Contracts\CategoryRepositoryInterface;
 use App\Http\Requests\CategoryRequest;
+use App\Models\Book;
 
 class CategoryRepository implements CategoryRepositoryInterface
 {
@@ -20,7 +21,7 @@ class CategoryRepository implements CategoryRepositoryInterface
 
     public function showCategory($id)
     {
-        return Category::where('id', '=', $id)->first();        
+        return Category::where('id', '=', $id)->first();
     }
 
     public function updateCategory(CategoryRequest $request, Category $category)
@@ -30,6 +31,12 @@ class CategoryRepository implements CategoryRepositoryInterface
 
     public function deleteCategory(Category $category)
     {
-        return $category->delete($category);
+        $books_with_the_category = Book::where('category_id', $category->id)->get();
+
+        if ($books_with_the_category !== null) {
+            return response()->json(['error' => 'Stop! Books are attached to this category'], 404);
+        } else {
+            return $category->delete($category);
+        }
     }
 }
